Actualizacion de las tablas para migracion

+++ db/migrate/20231001212724_create_clientes.rb
@@ -4,7 +4,7 @@ class CreateClientes < ActiveRecord::Migration[7.0]
       t.string :nombre, null: false
       t.text :direccion, default: "Sin dirección" 
       t.string :email, default: "Sin email" 
-      t.integer :telefono, null: false 
+      t.bigint :telefono, null: false 
 
       t.timestamps
     end
